// src/gen-ui/lib/core/base-types.ts
export enum Axis {
  horizontal = "horizontal",
  vertical = "vertical",
}
// ...
export enum WrapAlignment {
  start = "start",
  end = "end",
  center = "center",
  spaceBetween = "spaceBetween",
  spaceAround = "spaceAround",
  spaceEvenly = "spaceEvenly",
}
export enum WrapCrossAlignment {
  start = "start",
  end = "end",
  center = "center",
}